Add verifyHelpers.
